+++ lib/lightning_web/live/workflow_live/edit.ex
@@ -1421,6 +1421,7 @@ defmodule LightningWeb.WorkflowLive.Edit do
        page_title: "",
        selected_edge: nil,
        selected_job: nil,
+       last_selected_job: nil,
        selected_run: nil,
        selected_trigger: nil,
        selection_mode: nil,
@@ -3584,10 +3585,19 @@ defmodule LightningWeb.WorkflowLive.Edit do
   end
 
   defp assign_chat_session_id(socket, params) do
+    job_chat_session_id =
+      if changed?(socket, :selected_job) &&
+           not is_nil(socket.assigns[:last_selected_job]) do
+        nil
+      else
+        params["j-chat"]
+      end
+
     socket
     |> assign(
       workflow_chat_session_id: params["w-chat"],
-      job_chat_session_id: params["j-chat"]
+      job_chat_session_id: job_chat_session_id,
+      last_selected_job: socket.assigns[:selected_job]
     )
   end